allocation and editing
