+++ src/pages/Admin/Charity/Dashboard/Schedule/index.ts
@@ -1,0 +1,1 @@
+export { Schedule } from "./Schedule";
