--- src/pages/Admin/Charity/Dashboard/ScheduleEdit.tsx
@@ -1,1 +1,0 @@
-export function ScheduleEdit() {}
